# Batched Quantized Matmul + Fast Small QMV