- Support batched inputs for quantized matmuls
- Add a QMV specialization for `D=64,128`

The batched `qmv` and `qvm` will be useful for a quantized KV cache.

The existing `qmv` was pretty slow for `D=128` which is by far the most common attention head dim in LLMs.
This is because if you're using `simd_sum` for `D=128` you're only processing 4 values for each of the 32 threads in your simdgroup. This isn't enough to max out the memory bandwidth and since that's the main bound for `qmv`, you sacrifice a lot of performance.

Instead, the new specialization uses `quad_sum` over quadgroups of 4 threads. This means each thread processes 32 values and the overall throughput is much higher.

On M1 Max:
```
queries = (1, 8, 4, 1, 128)
keys = (1, 8, 4, 32768, 128)
```
Before:
```
Timing matmul ... 1.68047 msec
Timing quant_matmul ... 2.66873 msec
```
After:
```
Timing matmul ... 1.67618 msec
Timing quant_matmul ... 0.54323 msec
```