@barronalex 2024-10-18
Looks like we were slightly regressing performance so I added a `batched` template param.
Now seems like we're at parity with `main`:

On M2 Ultra with a 3B model
Before: `180.6` tks (avg of 3)
After:    `180.6` tks (avg of 3)